refactor: fixed app
