+++ app/components/index.js
@@ -2,5 +2,5 @@ export { default as AppBar } from './global/AppBar'
 export {
   default as AuthenticatedComponent,
 } from './global/AuthenticatedComponent'
-export { default as Dashboard } from './dashboard/Dashboard'
-export { default as LoginPage } from './login/LoginPage'
+export { default as Dashboard } from './pages/dashboard/Dashboard'
+export { default as LoginPage } from './pages/login/LoginPage'
