+++ app/components/pages/manuscript/Manuscript.js
