+++ app/components/pages/manuscript/Manuscript.md
