+++ app/components/pages/submission-wizard/SubmissionPage.js
@@ -1,14 +1,14 @@
 import React from 'react'
 import { Route, Switch } from 'react-router-dom'
 import WithCurrentSubmission from './WithCurrentSubmission'
-import AuthorDetailsPage from './AuthorDetails/AuthorDetailsPage'
-import FileUploadsPage from './FileUploads/FileUploadsPage'
-import ManuscriptMetadata from './ManuscriptMetadata/ManuscriptMetadata'
-import ReviewerSuggestions from './ReviewerSuggestions/ReviewerSuggestions'
-import { schema as fileUploadsSchema } from './FileUploads/FileUploadsSchema'
-import { schema as authorDetailsSchema } from './AuthorDetails/AuthorDetailsSchema'
-import { schema as manuscriptMetadataSchema } from './ManuscriptMetadata/ManuscriptMetadataSchema'
-import { schema as reviewerSuggestionsSchema } from './ReviewerSuggestions/ReviewerSuggestionsSchema'
+import AuthorDetailsPage from './Author/AuthorDetailsPage'
+import FileUploadsPage from './Files/FileUploadsPage'
+import ManuscriptMetadata from './Submission/ManuscriptMetadata'
+import ReviewerSuggestions from './Editors/ReviewerSuggestions'
+import { schema as fileUploadsSchema } from './Files/FileUploadsSchema'
+import { schema as authorDetailsSchema } from './Author/AuthorDetailsSchema'
+import { schema as manuscriptMetadataSchema } from './Submission/ManuscriptMetadataSchema'
+import { schema as reviewerSuggestionsSchema } from './Editors/ReviewerSuggestionsSchema'
 import WizardStep from './WizardStep'
 
 const SubmissionPage = ({ match, history }) => (
