+++ app/components/pages/submission-wizard/WizardStep.js
@@ -2,8 +2,8 @@ import React from 'react'
 import { Formik } from 'formik'
 import { Box, Flex } from 'grid-styled'
 import { Button } from '@pubsweet/ui'
-import ButtonLink from '../ui/atoms/ButtonLink'
-import { FormH2 } from '../ui/atoms/FormHeadings'
+import ButtonLink from '../../ui/atoms/ButtonLink'
+import { FormH2 } from '../../ui/atoms/FormHeadings'
 import AutoSave from './AutoSave'
 import ProgressBar from './ProgressBar'
